add files to editorContainer
