+++ src/client/app/action/editorContainer.jsx
@@ -106,3 +106,13 @@ export function setEditorSize(width, height) {
     });
   };
 }
+
+export function updateFile(index, content) {
+  return (dispatch) => {
+    dispatch({
+      type: ActionTypes.UPDATE_FILE,
+      index,
+      content
+    });
+  };
+}
