--- src/client/app/action/p5files.jsx
@@ -1,11 +1,0 @@
-import * as ActionTypes from '../constants.jsx';
-
-export function updateFile(index, content) {
-  return (dispatch) => {
-    dispatch({
-      type: ActionTypes.UPDATE_FILE,
-      index,
-      content
-    });
-  };
-}
